feat(UIFactory): extract defaultUIFactory to separate file [YTFRONT-3814]
We don't need default implementation of UIFactory in case of a custom version of UI.
So such splitting allows to get rid of unnecessary code from js-files.
