+++ packages/ui/src/ui/entries/main.tsx
@@ -6,11 +6,12 @@ import {handleAuthError} from '../store/actions/global';
 import {YT_UI_CLUSTER_HEADER_NAME} from '../../shared/constants';
 import './main.scss';
 import './yfm-overrides.scss';
+import {defaultUIFactory} from '../UIFactory/default-ui-factory';
 
 yt.subscribe('error', onError);
 axios.interceptors.response.use(undefined, onAxiosError);
 
-renderApp({});
+renderApp(defaultUIFactory);
 
 function onAxiosError(e: any) {
     onError(e);
